+++ CHANGELOG.md
@@ -11,10 +11,11 @@ These changes are available in the [master branch](https://github.com/PrefectHQ/
 ### Enhancements
 
 - Better exception for unsubscriptable mapping arguments - [#1821](https://github.com/PrefectHQ/prefect/issues/1821)
+- Add DaskGateway tip to docs - [#1959](https://github.com/PrefectHQ/prefect/issues/1959)
 
 ### Task Library
 
-- None
+- Fix `V1DeleteOptions` call for Kubernetes tasks - [#2050](https://github.com/PrefectHQ/prefect/pull/2050)
 
 ### Fixes
 
@@ -30,7 +31,7 @@ These changes are available in the [master branch](https://github.com/PrefectHQ/
 
 ### Contributors
 
-- None
+- [Romain Thalineau](https://github.com/romaintha)
 
 ## 0.9.4 <Badge text="beta" type="success"/>